Move Tap to Place out of Experimental into Solvers folders
