+++ Assets/MixedRealityToolkit.Examples/Demos/Solvers/Scenes/TapToPlaceExample.unity
