+++ Assets/MixedRealityToolkit.Examples/Demos/Solvers/Scripts/HideTapToPlaceLabel.cs.meta
