--- Assets/MixedRealityToolkit.Examples/Experimental/TapToPlace.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 46715b270fbe218439552f7253accd0d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
