--- Assets/MixedRealityToolkit.SDK/Experimental/TapToPlace.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 934bd4be613889847a778f3a8260ee24
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
