+++ Assets/MixedRealityToolkit.SDK/Features/Utilities/Solvers/TapToPlace.cs.meta
