--- Assets/MixedRealityToolkit.SDK/Inspectors/Experimental/TapToPlace.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9d280220445a7c141818ad27e7f9b2f7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
