+++ Assets/MixedRealityToolkit.SDK/Inspectors/Utilities/Solvers/TapToPlaceInspector.cs
@@ -1,10 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Experimental.Utilities;
 using UnityEditor;
 
-namespace Microsoft.MixedReality.Toolkit.Experimental.Inspectors
+namespace Microsoft.MixedReality.Toolkit.Utilities.Editor.Solvers
 {
     /// <summary>
     /// Custom inspector for the Tap to Place class
